Add timestamp for at DVH can read timesatmp

+++ src/main/kotlin/no/nav/syfo/oppfolgingsplanmottak/kafka/domain/KFollowUpPlan.kt
@@ -5,5 +5,5 @@ data class KFollowUpPlan(
     val fodselsnummer: String,
     val virksomhetsnummer: String,
     val behovForBistandFraNav: Boolean,
-    val opprettet: Int
+    val opprettet: Long
 )
